clean: Organize the directory.
